Adjust paths and create plots of diagnostic images
